backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlmodel[asyncio]
asyncpg 
alembic 
pydantic-settings 
python-jose[cryptography] 
sqlalchemy
httpx
fastfeedparser
jwt 
pydantic[email]
python-multipart
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
google-genai
pygooglenews
elevenlabs
google-cloud-storage